requirements.txt
absl-py==0.14.0
albumentations==1.1.0
astunparse==1.6.3
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///tmp/build/80754af9/bleach_1628110601003/work
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.6
clang==5.0
cycler==0.10.0
efficientnet-pytorch==0.6.3
entrypoints==0.3
fastcluster==1.1.26
ffmpy==0.2.3
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
jupyter==1.0.0
kaggle==1.5.12
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
mkl-service==2.4.0
munch==2.5.0
networkx==2.6.3
nltk==3.6.5
numpy==1.19.5
nvidia-ml-py==11.470.66
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.3.56
opencv-python-headless==4.5.4.58
opt-einsum==3.3.0
pandas==1.3.2
Pillow==8.3.2
pretrainedmodels==0.7.4
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-slugify==5.0.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==6.0
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.1
six==1.15.0
smp==0.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
terminado==0.9.4
text-unidecode==1.3
tifffile==2021.11.2
timm==0.4.12
torch==1.10.0
torchaudio==0.9.0
torchvision==0.11.1
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
